Update docs on value lifetime
The previous definition had an issue because the following could
happen:

1) A calls p := v.Get()
2) GOMAXPROCS lowers
3) B calls v.Do(), lifetime ends
4) A uses p

It seems in general, we'd need the API more complicated
if we wanted to be able to deallocate values.
Probably not worth it, if GOMAXPROCS gets lower, the extra values
shouldn't cause any harm, they just would not be used.